backend/requirements.txt
langchain>=0.0.267
langchain-google-genai>=0.0.5
chromadb>=0.4.13
pdfplumber>=0.10.2
python-dotenv>=1.0.0
streamlit>=1.26.0
pypdf>=3.15.1
sentence-transformers>=2.2.2
tiktoken>=0.5.1